Add first tests for FMIMapper and FMIBridge using some test FMUs from https://github.com/modelica/fmi-cross-check
Used FMUs from https://github.com/modelica/fmi-cross-check are licensed under 2-Clause BSD according to their License.txt files in the *.fmu
-  Feedthrough.fmu passes values from its input variables to its output variables on every step (with some quirk for the boolean, as testing showed)
-  Stair.fmu has a sole integer output named 'counter' that counts the seconds since t=0. Plotting 'counter' would show a Stair

Edited CMakleLists.txt to copy files relevant for testing (config.yaml, *.fmu) to the build folder as part of the test target (test_CoSiMa)

+++ src/Mapper/FMIMapper.cpp
@@ -69,7 +69,7 @@ int FMIMapper::readConfiguration(configVariants_t configVariants) {
 					}
 					else {
 						config.floatInputList.push_back(NamesAndIndex(modelDescription->model_identifier + "." + var.name, var.name, (int)state->floats.size()));
-						state->floats.push_back(realVar.start().value());
+						state->floats.push_back((float)realVar.start().value());
 					}
 				}
 				else {
